Move PasswordHasher and RandomGenerator into components
