+++ src/FunderMaps.AspNetCore/Helpers/JwtHelper.cs
@@ -23,7 +23,7 @@ namespace FunderMaps.AspNetCore.Helpers
 
             // FUTURE: Write warning to log
             // FUTURE: Using IRandom
-            using var rng = new Core.Services.RandomGenerator();
+            using var rng = new Core.Components.RandomGenerator();
             return new SymmetricSecurityKey(rng.GetRandomByteArray(32));
         }
     }
